pyproject.toml
[tool.poetry]
name = "boardie-backoffice-backend"
version = "0.1.0"
description = "HTTP API for the Boardie Backoffice"
authors = ["Alex Brou <alexandrepilarbrou@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {version = "0.87.0",  extras = ["all"]}
SQLAlchemy = {version = "1.4.44", extras = ["mypy"]}
alembic = "1.9.0"
asyncpg= "0.27.0"
greenlet = ">=2.0.1"
pyjwt = "^2.6.0"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^5.0.4"
mypy = "^0.982"
poethepoet = "^0.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poe.tasks]
test = "pytest ./tests"
check_fmt = "black --check ./api"
check_imports = "isort --check ./api"
check_lint = "flake8 ./api"
check_types = "mypy ./api"
check_all = ["check_fmt", "check_imports", "check_lint", "check_types"]
check_and_test = ["check_fmt", "check_imports", "check_lint", "check_types", "test"]